welcome page created, visitor registration moved to new template, mapping updated
