+++ src/main/java/com/advcourse/conferenceassistant/controller/ErrorPageController.java
@@ -4,10 +4,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class ErrorPageContrroller {
+public class ErrorPageController {
 
     @GetMapping("/forbidden")
-    public String getFrobidenPage(){
+    public String getForbiddenPage(){
         return "forbidden-page";
     }
 }
