+++ src/main/java/com/advcourse/conferenceassistant/controller/TempControler.java
@@ -12,5 +12,4 @@ public class TempControler {
         return "schedule";
     }
 
-
 }
