+++ src/main/java/com/advcourse/conferenceassistant/service/impl/TopicServiceImpl.java
@@ -7,6 +7,7 @@ import com.advcourse.conferenceassistant.service.dto.TopicDto;
 import com.advcourse.conferenceassistant.service.dto.mapper.TopicMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -55,9 +56,11 @@ public class TopicServiceImpl implements TopicService {
         topic.setStart(dto.getStart());
         topic.setTheme(dto.getTheme());
         topic.setConfId(dto.getConfId());
+
         return save(topic);
     }
     @Override
+    @Nullable
     public TopicDto findActiveTopicByConfId(long confId) {
         List<Topic> byConferenceId = topicRepository.findByConferenceId(confId);
         List<TopicDto> topics = TopicMapper.toDtos(byConferenceId);
@@ -67,7 +70,7 @@ public class TopicServiceImpl implements TopicService {
                 return t;
             }
 //            TODO: avoid this null returning if possible
-        } log.info("No active topic found");
+        } log.debug("No active topic found");
         return null;
     }
